Update next config
