--- apps/dataset-browser/next.config.js
@@ -1,13 +1,0 @@
-/** @type {import('next').NextConfig} */
-
-const withNextIntl = require('next-intl/plugin')('./src/i18n.ts');
-const withMDX = require('@next/mdx')();
-
-const nextConfig = {
-  transpilePackages: ['ui'],
-  experimental: {
-    mdxRs: true,
-  },
-};
-
-module.exports = withNextIntl(withMDX(nextConfig));
