+++ apps/dataset-browser/next.config.mjs
@@ -1,0 +1,17 @@
+/**
+ * @type {import('next').NextConfig}
+ */
+
+import NextIntlPlugin from 'next-intl/plugin';
+import MDXPlugin from '@next/mdx';
+const withNextIntl = NextIntlPlugin('./src/i18n.ts');
+const withMDX = MDXPlugin();
+
+const nextConfig = {
+  transpilePackages: ['ui'],
+  experimental: {
+    mdxRs: true,
+  },
+};
+
+export default withNextIntl(withMDX(nextConfig));
